First code for JCfg project.
This project aims to be a library/module to provide a pythonic and
object oriented interface to Juniper Networks device configurations.

At this point, there are some of the basics of how I hope for this
project to come together...it may even have enough to be somewhat useful
for very limited use cases.  There's a lot of work to be done, here, but
so far I'm happy with how it is coming together.

JCfg makes use of Juniper's excellent pyez project for some of the
underlying communication to and from JunOS devices. (Hat tip to the
various folks that have been involved in creating and improving that
library over the years.  Jeremy, don't forget, bourbon is on me if you
ever get back to Louisville.)  By way of pyez, the interface to the
config is by way of lxml's ETree API.  It is the difficulty of dealing
with the XML representation of the configuration that inspired me to
develop a more pythonic and object oriented approach to interfacing with
JunOS configs.

Please be aware that I am a relative n00b when it comes to writing
python code...having never built and run a major python project...or a
major project in any language, to be honest.  I will make mistakes and
demonstrate a distinct lack of understanding about things that
experienced software developers take for granted.  Bear with me.