Refactoring and renameing classes.
A slight change in direction to remove "CFG" from the classes...I want
the classes to represent both config and active status of the thing it's
modeling, not specifically config, or separately status (it was quickly
looking like it was going to be hard to manage separate coordinated
classes).

It's still, in my mind, extremely early days for this project, so don't
count on APIs to be at all stable right now.