Add basic code for JunosDev.commit_confirm()
Totally single-threaded and synchronous...it will issue a commit confirm
using rollback for the confirmation timeout, delay for the delay time
period, and then issue a commit.

rollback defaults to 10 minutes
delay defaults to 2 minutes

Let me re-iterate that this blocks for the time period specified in
(minutes) in delay (default: 2), so this isn't going to end well if
you're trying to use this to configured lots of devices in a script.

The idea behind this method is to use for changes that you are worried
might cause you to lose contact with the device (the same reason I
pretty incessantly use "commit confirm" when I'm manually configuring
devices).  In my experience, waiting 2 minutes is typically long enough
for whatever changes are going to occur as a result of routing protocol
failures (such as iBGP sessions losing connectivity, that could then
affect your connectivity to the device) to actually occur so the impact
of the changes are felt.  If you can still talk to the device to issue
a full commit after 2 minutes, in my experience, you're usually in good
shape.  The method takes care of that delay and issuing the full commit
for you.  If the change causes the device to be unreachable, the full
commit will fail and the confirm timeout will eventually expire and the
normal rollback and hopefully restoration of management access will
occur.