Add rudimentary functionality for commit_check() in JunosDev
Load the current configuration state to the device (with overwrite),
then issue a "commit check".  If the config state is invalid, you'll get
an Exception.

Did a rollback in a finally block so that we return the device back to
it's active config and don't leave it in the current GinPy state for the
device as, if the check fails, it's an invalid config, and all of the
GinPy commit operations (at least thus far) assume that they will push
the new candidate config and retrieve_config() will currently retrieve
the candidate config.  Besides, it's just bad mojo to leave an
uncommitted, candidate config in place when you're done working with a
device.  (Kindergarten lesson: clean up after yourself)